refactor(agent): co-locate tool input schemas with their consumers
