+++ packages/agent/src/tools/dispatch.ts
@@ -1,3 +1,4 @@
+import { z } from "zod";
 import type { Tool } from "@openomni/protocol";
 import { IngressEngine } from "../ingress/engine";
 import {
@@ -9,7 +10,28 @@ import {
   type DispatchReviewInput,
   type DispatchTask,
 } from "../loop/execution-supervisor";
-import { DispatchInput } from "./schemas";
+
+export const DispatchInput = z.object({
+  objective: z.string().describe("Overall goal for task dispatch"),
+  tasks: z
+    .array(
+      z.object({
+        id: z.string().describe("Unique task identifier"),
+        description: z.string().describe("Task description"),
+        agentType: z.string().describe("Agent type to execute task"),
+        dependencies: z
+          .array(z.string())
+          .default([])
+          .describe("Task IDs this task depends on"),
+        fileScope: z
+          .array(z.string())
+          .default([])
+          .describe("Files this task operates on"),
+      }),
+    )
+    .describe("Array of tasks to dispatch"),
+});
+export type DispatchInput = z.infer<typeof DispatchInput>;
 
 export type { DispatchContext, DispatchReviewDecision, DispatchReviewInput };
