+++ packages/agent/src/tools/index.ts
@@ -1,10 +1,9 @@
-export { SubagentInput, DispatchInput, ScheduleInput } from "./schemas";
-export { Dispatch } from "./dispatch";
+export { SubagentInput, Subagent } from "./subagent";
+export type { SubagentContext, SubagentResult } from "./subagent";
+export { DispatchInput, Dispatch } from "./dispatch";
 export type {
   DispatchContext,
   DispatchReviewDecision,
   DispatchReviewInput,
 } from "./dispatch";
-export { ScheduleTool } from "./schedule";
-export { Subagent } from "./subagent";
-export type { SubagentContext, SubagentResult } from "./subagent";
+export { ScheduleInput, ScheduleTool } from "./schedule";
